+++ test/integration/spec/remotetracks.js
@@ -43,7 +43,7 @@ describe('RemoteParticipant', function() {
       [, thisRoom, thoseRooms] = await setup({
         testOptions: {
           bandwidthProfile: {
-            video: { maxTracks: 1 }
+            video: { maxTracks: 1,  dominantSpeakerPriority: 'low' }
           },
           tracks: [dataTrack]
         },
@@ -108,7 +108,7 @@ describe('RemoteParticipant', function() {
 
     // eslint-disable-next-line no-warning-comments
     // TODO: enable these tests when track_priority MSP is available in prod
-    if (defaults.topology !== 'peer-to-peer' && defaults.environment === 'dev') {
+    if (defaults.topology !== 'peer-to-peer' && defaults.environment === 'stage') {
       it('subscriber can upgrade track\'s effective priority', async () => {
         await waitFor([
           trackSwitchedOn(bobRemoteVideoTrack),
@@ -126,10 +126,7 @@ describe('RemoteParticipant', function() {
         ], 'Alice track to get switched On, and Bob Switched Off');
       });
 
-      // eslint-disable-next-line no-warning-comments
-      // TODO: can subscriber downgrade tracks priority?
-      // https://twilio.slack.com/archives/C8HJ77RJM/p1568842297013800?thread_ts=1568405331.070800&cid=C8HJ77RJM
-      it.skip('subscriber can downgrade track\'s effective priority', async () => {
+      it('subscriber can downgrade track\'s effective priority', async () => {
         await waitFor([
           trackSwitchedOn(bobRemoteVideoTrack),
           trackSwitchedOff(aliceRemoteVideoTrack)